Cambio de tipo Integer a Long en las interfaces CRUD de Merchant y Client

+++ src/main/java/backend/merchants/MerchantRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MerchantRepository extends CrudRepository<Merchant, Integer> {
+public interface MerchantRepository extends CrudRepository<Merchant, Long> {
 }
